Implement hints for python runner
