+++ cairo/tests/programs/test_os.cairo
@@ -1,0 +1,7 @@
+from programs.os import main
+
+func test_os{output_ptr: felt*}() {
+    main();
+
+    return ();
+}
