+++ cairo/tests/programs/test_os.py
@@ -1,0 +1,4 @@
+class TestOs:
+
+    def test_os(self, cairo_run):
+        assert cairo_run("test_os") == list(range(1, 9))
