+++ cairo/tests/utils/hints.py
@@ -32,6 +32,40 @@ def new_default_dict(
     return base
 
 
+block_info = """
+ids.block_info.coinbase = 1
+ids.block_info.timestamp = 2
+ids.block_info.number = 3
+ids.block_info.prev_randao.low = 4
+ids.block_info.prev_randao.high = 5
+ids.block_info.gas_limit = 6
+ids.block_info.chain_id = 7
+ids.block_info.base_fee = 8
+"""
+
+
+transaction_hash = """
+"""
+
+hints = {"block_info": block_info, "transaction_hash": transaction_hash}
+
+
+def implement_hints(program):
+    return {
+        k: [
+            (
+                CairoHint(
+                    accessible_scopes=hint_.accessible_scopes,
+                    flow_tracking_data=hint_.flow_tracking_data,
+                    code=hints.get(hint_.code, hint_.code),
+                )
+            )
+            for hint_ in v
+        ]
+        for k, v in program.hints.items()
+    }
+
+
 @contextmanager
 def patch_hint(program, hint, new_hint):
     patched_hints = {
